update test classes

+++ src/test/java/com/webapp/portfolio/PortfolioRepositoryTest.java
@@ -3,8 +3,10 @@ package com.webapp.portfolio;
 import com.webapp.user.Role;
 import com.webapp.user.User;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -27,15 +29,15 @@ class PortfolioRepositoryTest {
                 .ticker("AAPL")
                 .company("Apple Inc.")
                 .sharesBuy(100L)
-                .buy(150.0)
-                .sell(175.0)
-                .sharesSell(50)
-                .interestBuySell(25.0)
-                .profitLossBuySell(125.0)
-                .dividend(1.0)
-                .interestDividend(0.25)
-                .profitLossDividend(0.75)
-                .sumProfitLoss(125.75)
+                .buy(BigDecimal.valueOf(150.0))
+                .sell(BigDecimal.valueOf(175.0))
+                .sharesSell(50L)
+                .interestBuySell(BigDecimal.valueOf(25.0))
+                .profitLossBuySell(BigDecimal.valueOf(125.0))
+                .dividend(BigDecimal.valueOf(1.0))
+                .interestDividend(BigDecimal.valueOf(0.25))
+                .profitLossDividend(BigDecimal.valueOf(0.75))
+                .sumProfitLoss(BigDecimal.valueOf(125.75))
                 .build();
         portfolioRepository.save(portfolio);
         List<Portfolio> portfolios = portfolioRepository.findByUserId(user.getId());
